Ismael's comments

+++ streams/src/main/java/org/apache/kafka/streams/state/internals/OffsetCheckpoint.java
@@ -31,7 +31,6 @@ import java.io.InputStreamReader;
 import java.io.OutputStreamWriter;
 import java.nio.charset.StandardCharsets;
 import java.nio.file.Files;
-import java.nio.file.NoSuchFileException;
 import java.util.Collections;
 import java.util.HashMap;
 import java.util.Map;
@@ -171,9 +170,9 @@ public class OffsetCheckpoint {
      * @throws IOException if there is any IO exception during delete
      */
     public void delete() throws IOException {
-        try {
+        if (Files.exists(file.toPath())) {
             Files.delete(file.toPath());
-        } catch (final NoSuchFileException e) { }
+        }
     }
 
     @Override
